refactor(map): story_map_panel.dart 핀 마커 + 순수 함수 추출
큰 파일을 leaf부터 분해. 카메라 애니메이션 로직(state 의존성 큼)은
유지하고, presentational/순수 부분만 안전하게 추출.

신설:
- lib/widgets/map/pin_marker.dart (~170줄, part 파일)
  · _PinStyle, _MarkerNode, _IterableX
  · _CompactPinMarker, _PinNumberBadge, _PinPointerPainter
  · 모두 외부 state 의존 없는 presentational 위젯/데이터 클래스

- lib/utils/map_math.dart (~130줄, 공개 함수)
  · easeInOut: 4차 ease-in-out 보간
  · mercatorY: Web Mercator Y 좌표 투영
  · normalizedLongitudeDelta: 경도 차이 보정
  · hasMultiPlacePin / splitPlaceParts / buildSplitPinPoints: 다중 장소 핀
  · buildAdjustedPoints: 가까운 좌표 분산 배치
  · rotateOffset: 화면 좌표 회전
  · eventListSignature: 이벤트 리스트 변경 감지

- test/utils/map_math_test.dart (27 tests)
  · 모든 순수 함수에 대한 단위 테스트

부모 파일:
- 인라인 정의 제거, map_math.xxx() 형태로 호출
- part 'map/pin_marker.dart' 디렉티브 추가
- 1500 → 1222줄 (-19%)

flutter analyze: 0 issues
flutter test: 107 tests pass (80 + 27 신규)
