+++ lib/widgets/weekly_tab_page.dart
@@ -881,5 +881,4 @@ class _WeeklyTabPageState extends ConsumerState<WeeklyTabPage> {
       ),
     );
   }
-
 }
